Refactoring
- Renamed and moved the item interfaces
