+++ app/Interfaces/Item/IModel.php
@@ -1,15 +1,41 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Interfaces;
+namespace App\Interfaces\Item;
 
-interface ResourceTypeItemModel
+use Illuminate\Database\Eloquent\Model;
+
+interface IModel
 {
+    /**
+     * Return an instance of the relevant item model
+     *
+     * @param int $item_id
+     *
+     * @return Model|null
+     */
+    public function instance(int $item_id): ?Model;
+
+    /**
+     * Take the Item and ItemType model and return a formatted array for
+     * output
+     *
+     * @param Model $item
+     * @param Model $item_type
+     *
+     * @return array
+     */
+    public function instanceToArray(
+        Model $item,
+        Model $item_type
+    ): array;
+
     /**
      * Return an array of the item results based on the requested
      * parameters
      *
      * @param int $resource_type_id
+     * @param int $resource_id
      * @param int $offset
      * @param int $limit
      * @param array $parameters
@@ -21,6 +47,7 @@ interface ResourceTypeItemModel
      */
     public function paginatedCollection(
         int $resource_type_id,
+        int $resource_id,
         int $offset = 0,
         int $limit = 10,
         array $parameters = [],
@@ -30,9 +57,25 @@ interface ResourceTypeItemModel
     ): array;
 
     /**
+     * Return a single item as an array
+     *
+     * @param int $resource_type_id
+     * @param int $resource_id
+     * @param int $item_id
+     *
+     * @return array|null
+     */
+    public function single(
+        int $resource_type_id,
+        int $resource_id,
+        int $item_id
+    ): ?array;
+
+    /**
      * Return the total number of items that match the requested parameters
      *
      * @param int $resource_type_id
+     * @param int $resource_id
      * @param array $parameters
      * @param array $search_parameters
      * @param array $filter_parameters
@@ -41,6 +84,7 @@ interface ResourceTypeItemModel
      */
     public function totalCount(
         int $resource_type_id,
+        int $resource_id,
         array $parameters = [],
         array $search_parameters = [],
         array $filter_parameters = []
