+++ app/Item/SimpleItem.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace App\Item;
 
+use App\Interfaces\Item\IModel;
 use App\Models\ItemType\SimpleItem as ItemModel;
 use App\Models\Transformers\Transformer;
 use App\Validators\Fields\ItemType\SimpleItem as ItemTypeSimpleItemValidator;
@@ -66,9 +67,9 @@ class SimpleItem extends AbstractItem
     /**
      * Return the model instance for the item type
      *
-     * @return \App\Interfaces\ItemModel
+     * @return IModel
      */
-    public function model(): \App\Interfaces\ItemModel
+    public function model(): IModel
     {
         return new ItemModel();
     }
