+++ app/Models/ItemType/AllocatedExpense.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Models\ItemType;
 
-use App\Interfaces\ItemModel;
+use App\Interfaces\Item\IModel;
 use App\Utilities\General;
 use App\Utilities\Model as ModelUtility;
 use Illuminate\Database\Query\Builder as QueryBuilder;
@@ -18,7 +18,7 @@ use Illuminate\Support\Facades\DB;
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class AllocatedExpense extends Model implements ItemModel
+class AllocatedExpense extends Model implements IModel
 {
     protected $table = 'item_type_allocated_expense';
