+++ app/Models/ItemType/SimpleItem.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Models\ItemType;
 
-use App\Interfaces\ItemModel;
+use App\Interfaces\Item\IModel;
 use App\Utilities\Model as ModelUtility;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use Illuminate\Database\Eloquent\Model;
@@ -16,7 +16,7 @@ use Illuminate\Database\Eloquent\Model;
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class SimpleItem extends Model implements ItemModel
+class SimpleItem extends Model implements IModel
 {
     protected $table = 'item_type_simple_item';
