+++ app/ResourceTypeItem/SimpleExpense.php
@@ -3,10 +3,9 @@ declare(strict_types=1);
 
 namespace App\ResourceTypeItem;
 
-use App\Interfaces\ResourceTypeItemModel;
+use App\Interfaces\ResourceTypeItem\IModel;
 use App\Models\ResourceTypeItemType\SimpleExpense as ItemModel;
 use App\Models\Transformers\Transformer;
-use Illuminate\Database\Eloquent\Model;
 
 /**
  * The Interface for dealing with simple expenses, everything should be
@@ -67,9 +66,9 @@ class SimpleExpense extends AbstractItem
     /**
      * Return the model instance for resource type item type
      *
-     * @return ResourceTypeItemModel
+     * @return IModel
      */
-    public function model(): ResourceTypeItemModel
+    public function model(): IModel
     {
         return new ItemModel();
     }
